Display images in reverse, major refactor and cleanup
