+++ package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "chalk": "^4.1.2",
     "inquirer": "^12.7.0",
+    "jimp": "^1.6.0",
     "terminal-image": "^3.1.1"
   }
 }
